school/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2024.10
cssselect2==0.7.0
Django==5.1.1
django-crispy-forms==2.3
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
docopt==0.6.2
fontawesomefree==6.6.0
fonttools==4.54.1
num2words==0.5.13
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.1
django-admin-tools==0.9.3
django-crispy-forms==2.3
django-grappelli==4.0.1
django-simple-history==3.7.0
django-tailwind==3.8.0
django-webpush==0.3.6
django-widget-tweaks==1.5.0
docopt==0.6.2
fontawesomefree==6.6.0
fonttools==4.54.1
frozenlist==1.5.0
http_ece==1.2.1
idna==3.10
multidict==6.1.0
num2words==0.5.13
pillow==10.4.0
propcache==0.2.1
py-vapid==1.9.2
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
pywebpush==2.0.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
